requirements.txt
numpy==1.25.2
pandas==2.0.3
Pillow==10.0.0
Pillow==10.0.0
scikit_learn==1.3.0
scipy==1.11.2
statsmodels==0.14.0
torch==2.0.1
torchvision==0.15.2
